requirements.txt
sqlalchemy
python-dateutil
pillow
requests
numpy
pandas
matplotlib